+++ packages/node/src/methods/app-instance-operations.ts
@@ -2,7 +2,7 @@ import { Node } from "@counterfactual/common-types";
 
 import { Channels } from "../channels";
 import { APP_INSTANCE_STATUS } from "../models";
-import { IMessagingService } from "../service-interfaces";
+import { IMessagingService } from "../services";
 
 export async function getInstalledAppInstances(
   channels: Channels,
