+++ packages/node/src/methods/multisig-operations.ts
@@ -2,7 +2,7 @@ import { Node } from "@counterfactual/common-types";
 
 import { Channels } from "../channels";
 import { NodeMessage } from "../node";
-import { IMessagingService } from "../service-interfaces";
+import { IMessagingService } from "../services";
 
 /**
  * This creates a multisig while sending details about this multisig
